Feat: Basic context menu handler.
Other programs, like Jetbrains IDEs or 7zip, use some CLSID-like format to give handlers their own text and icons. This is a temporary measure until I learn about that. For some reason, the shell will never launch the actual .ps1 file, even if in the PATH or if given the absolute path.